moved test file to top level tests directory
